+++ src/sensors/fake_temperature.hpp
@@ -28,7 +28,7 @@ class FakeTemperature : public TemperatureInterface {
    *
    * @return int temperature degrees C
    */
-  int getData() override;
+  uint8_t getData() override;
 
   /**
    * @brief waits for acceleration, generate random time for failure
